consolidate structure.go
